twhile type debug

+++ libs/compiler/codegen.c
@@ -406,14 +406,14 @@ void Stmt_gen(syntax *const sx, ad *const context)
 		}
 		case TWhile:
 		{
-			int oldbreak = context->adbreak;
-			int oldcont = context->adcont;
-			int ad = (int)mem_get_size(sx);
+			size_t oldbreak = context->adbreak;
+			size_t oldcont = context->adcont;
+			size_t ad = mem_get_size(sx);
 
 			context->adcont = ad;
 			Expr_gen(sx, 0);
 			tocode(sx, BE0);
-			context->adbreak = (int)mem_get_size(sx);
+			context->adbreak = mem_get_size(sx);
 			mem_add(sx, 0);	
 			Stmt_gen(sx, context);
 			adcontbeg(sx, context, ad);
